requirement.txt
sentence_transformers
tqdm
faiss-cpu
httpx